--- a/ci/actions/windows/disable_windows_defender.ps1
+++ b/ci/actions/windows/disable_windows_defender.ps1
@@ -1,0 +1,5 @@
+$ErrorActionPreference = "Continue"
+
+Set-MpPreference -DisableArchiveScanning $true
+Set-MpPreference -DisableRealtimeMonitoring $true
+Set-MpPreference -DisableBehaviorMonitoring $true