--- a/ci/actions/windows/signing.ps1
+++ b/ci/actions/windows/signing.ps1
@@ -1,0 +1,9 @@
+$ErrorActionPreference = "Continue"
+
+if (Test-Path env:CSC_LINK) {
+  $path = Join-Path -Path "$env:TMP" -ChildPath csc.p12
+  [IO.File]::WriteAllBytes($path, [Convert]::FromBase64String($env:CSC_LINK))
+  $args = -split 'sign /a /ph /tr http://timestamp.digicert.com /td sha256 /fd sha256'
+  $args += @('/f', $path, '/p', $env:CSC_KEY_PASSWORD, "$env:BUILD_TYPE\*.exe")
+  . "C:\Program Files (x86)\Windows Kits\10\App Certification Kit\signtool.exe" $args
+}